Pantalla Analisis

+++ Inspinia_MVC5_SeedProject/Models/RequerimientoDetalle.cs
@@ -14,7 +14,7 @@ namespace PissanoApp.Models
 
         public int materialId { get; set; }
         public virtual Material Material { get; set; }
-        public int cantidad { get; set; }
+        public double cantidad { get; set; }
 
         public int partidaId { get; set; }
         public virtual Partida Partida { get; set; }
